+++ packages/fuselage-ui-kit/src/index.ts
@@ -2,4 +2,4 @@ export * from './hooks/useUiKitState';
 export * from './contexts/UiKitContext';
 export * from './surfaces';
 export { UiKitComponent } from './utils/UiKitComponent';
-export { extractInitialStateFromLayout } from './extractInitialStateFromLayout';
+export { extractInitialStateFromLayout } from './utils/extractInitialStateFromLayout';
